- refactor ledgerStore #137 - fix ConfirmationDialog for ledger (goBack) - fix transaction entry key for unconfirmed txes
