+++ package.json
@@ -10,6 +10,7 @@
     "@ledgerhq/hw-transport-u2f": "^4.48.0",
     "@material-ui/core": "^3.7.0",
     "@material-ui/icons": "^3.0.1",
+    "async-mutex": "^0.1.3",
     "babel-cli": "^6.26.0",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-preset-es2015": "^6.24.1",
@@ -20,7 +21,7 @@
     "delay": "^4.1.0",
     "downshift": "^3.1.7",
     "dpos-api-wrapper": "^1.3.2",
-    "dpos-ledger-api": "^2.0.0",
+    "dpos-ledger-api": "^3.0.1",
     "dpos-offline": "2.0.2",
     "inobounce": "^0.1.6",
     "is-mobile": "^2.0.0",
